+++ tests/testthat/test-docx.R
@@ -939,6 +939,27 @@ test_that("docs25: Label row is one cell.", {
   
 })
 
+test_that("docx26: Blank after on invisible column.", {
+  
+  fp <- file.path(base_path, "docx/test26.docx")
+  
+  tbl <- create_table(iris, borders = "all") %>%
+    define(Species, blank_after = TRUE, visible = FALSE)
+  
+  rpt <- create_report(fp, output_type = "DOCX") %>%
+    page_header("Left", "Right") %>%
+    add_content(tbl) %>%
+    page_footer("left", "", "right") %>%
+    titles("Table 1.0", "IRIS Data Frame",
+           blank_row = "below") %>%
+    footnotes("Here is a footnote", "And another")
+  
+  
+  res <- write_report(rpt)
+  
+  expect_equal(file.exists(fp), TRUE)
+  
+})