+++ tests/testthat/test-html.R
@@ -751,6 +751,30 @@ test_that("html22: Label row is one cell.", {
 })
 
 
+test_that("html23: Blank after on invisible column.", {
+  
+  fp <- file.path(base_path, "html/test23.html")
+  
+  tbl <- create_table(iris, borders = "all") %>%
+    define(Species, blank_after = TRUE, visible = FALSE)
+  
+  rpt <- create_report(fp, output_type = "HTML") %>%
+    page_header("Left", "Right") %>%
+    add_content(tbl) %>%
+    page_footer("left", "", "right") %>%
+    titles("Table 1.0", "IRIS Data Frame",
+           blank_row = "below") %>%
+    footnotes("Here is a footnote", "And another")
+  
+  
+  res <- write_report(rpt)
+  
+  expect_equal(file.exists(fp), TRUE)
+  
+})
+
+
+
 # User Tests --------------------------------------------------------------