+++ tests/testthat/test-pdf2.R
@@ -2118,6 +2118,27 @@ test_that("pdf2-55: Label row is one cell.", {
   
 })
 
+test_that("pdf2-56: Blank after on invisible column.", {
+  
+  fp <- file.path(base_path, "pdf2/test56.pdf")
+  
+  tbl <- create_table(iris, borders = "all") %>%
+    define(Species, blank_after = TRUE, visible = FALSE)
+  
+  rpt <- create_report(fp, output_type = "PDF", font = "Courier") %>%
+    page_header("Left", "Right") %>%
+    add_content(tbl) %>%
+    page_footer("left", "", "right") %>%
+    titles("Table 1.0", "IRIS Data Frame",
+           blank_row = "below") %>%
+    footnotes("Here is a footnote", "And another")
+  
+  
+  res <- write_report(rpt)
+  
+  expect_equal(file.exists(fp), TRUE)
+  
+})
 
 
 # # User Tests --------------------------------------------------------------