+++ tests/testthat/test-rtf.R
@@ -1280,3 +1280,25 @@ test_that("rtf38: RTF Image file works as expected.", {
     expect_equal(TRUE, TRUE)
 })
 
+test_that("rtf39: Blank after on invisible column.", {
+  
+  fp <- file.path(base_path, "rtf/test39.rtf")
+  
+  tbl <- create_table(iris, borders = "all") %>%
+    define(Species, blank_after = TRUE, visible = FALSE)
+  
+  rpt <- create_report(fp, output_type = "RTF") %>%
+    page_header("Left", "Right") %>%
+    add_content(tbl) %>%
+    page_footer("left", "", "right") %>%
+    titles("Table 1.0", "IRIS Data Frame",
+           blank_row = "below") %>%
+    footnotes("Here is a footnote", "And another")
+  
+  
+  res <- write_report(rpt)
+  
+  expect_equal(file.exists(fp), TRUE)
+  
+})
+