+++ tests/testthat/test-rtf2.R
@@ -2044,6 +2044,28 @@ test_that("rtf2-58: Label row is one cell.", {
   
 })
 
+test_that("rtf2-59: Blank after on invisible column.", {
+  
+  fp <- file.path(base_path, "rtf2/test59.rtf")
+  
+  tbl <- create_table(iris, borders = "all") %>%
+    define(Species, blank_after = TRUE, visible = FALSE)
+  
+  rpt <- create_report(fp, output_type = "RTF", font = "Courier") %>%
+    page_header("Left", "Right") %>%
+    add_content(tbl) %>%
+    page_footer("left", "", "right") %>%
+    titles("Table 1.0", "IRIS Data Frame",
+           blank_row = "below") %>%
+    footnotes("Here is a footnote", "And another")
+  
+  
+  res <- write_report(rpt)
+  
+  expect_equal(file.exists(fp), TRUE)
+  
+})
+
 # User Tests --------------------------------------------------------------
 
 test_that("rtf2-user1: demo table works.", {